test/ops_replayer: move loading files to ops_parser.cc
