+++ src/test/objectstore/allocsim/ops_parser.cc
@@ -2,6 +2,11 @@
 #include <cassert>
 #include <fmt/format.h>
 #include <iostream>
+#include <thread>
+#include <map>
+#include <sys/mman.h>
+#include <sys/stat.h>
+#include <fcntl.h>
 
 using namespace std;
 
@@ -164,5 +169,68 @@ void parse_entry_point(
     
     op_action(Op(at, ot, offset, length, object_ptr, collection_ptr, who_ptr));
     //context->ops.push_back(Op(at, ot, offset, length, object_ptr, collection_ptr, who_ptr));
+  }
+}
+
+
+
+void parse_files(
+  const vector<string>& input_files,
+  int nparser_threads,
+  vector<Op>& ops,
+  uint64_t& max_buffer_size)
+{
+  size_t sum_ops = 0;
+  for (auto &file : input_files) {
+    // Parse input file
+    vector<std::thread> parser_threads;
+    vector<shared_ptr<ParserContext>> parser_contexts;
+    cout << fmt::format("parsing file {}", file) << endl;
+    int fd = open(file.c_str(), O_RDONLY);
+    if (fd == -1) {
+        cout << "Error opening file" << endl;
+        exit(EXIT_FAILURE);
+    }
+    struct stat file_stat;
+    fstat(fd, &file_stat);
+    char* mapped_buffer = (char*)mmap(NULL, file_stat.st_size, PROT_READ, MAP_SHARED, fd, 0);
+    if (mapped_buffer == nullptr) {
+        cout << "error mapping buffer" << endl;
+        exit(EXIT_FAILURE);
+    }
+    uint64_t step_size = file_stat.st_size / nparser_threads;
+    char* start = mapped_buffer;
+    for (int i = 0; i < nparser_threads; i++) {
+      char* end = start + step_size;
+      if (i == nparser_threads - 1) {
+          end = mapped_buffer + file_stat.st_size;
+      }
+      while(*(end - 1) != '\n') {
+          end--;
+      }
+      shared_ptr<ParserContext> context = make_shared<ParserContext>();
+      context->start = start;
+      context->end = end;
+      context->max_buffer_size = 0;
+      parser_contexts.push_back(context);
+      auto op_action = [context](Op&& op) {
+        context->ops.push_back(op);
+      };
+      parser_threads.push_back(std::thread(parse_entry_point, context, op_action));
+      start = end;
+    }
+    for (auto& t : parser_threads) {
+        t.join();
+    }
+    // reduce
+    for (auto context : parser_contexts) {
+      sum_ops += context->ops.size();
+    }
+    ops.reserve(sum_ops);
+    for (auto context : parser_contexts) {
+        ops.insert(ops.end(), context->ops.begin(), context->ops.end());
+        max_buffer_size = max(context->max_buffer_size, max_buffer_size);
+    }
+    munmap(mapped_buffer, file_stat.st_size);
   }
 }
